Remove old notifications
