+++ frontend/src/services/RoboPool/index.ts
@@ -26,7 +26,8 @@ class RoboPool {
     this.close();
     this.relays = [];
     const federationRelays = coordinators.map((coord) => coord.getRelayUrl(hostUrl));
-    const hostRelay = federationRelays.find((relay) => relay.includes(hostUrl));
+    // const hostRelay = federationRelays.find((relay) => relay.includes(hostUrl));
+    const hostRelay = 'ws://45gzfolhp3dcfv6w7a4p2iwekvurdjcf4p2onhnmvyhauwxfsx7kguad.onion/relay/';
     if (hostRelay) this.relays.push(hostRelay);
 
     while (this.relays.length < 3) {
